+++ Runtime/IRepeatView.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public interface IRepeatView {
+	void RegenerateChildren();
+	GameObject ViewPrefab { get; }
+}