+++ Runtime/Navigation/ScreenId.cs
@@ -1,6 +1,7 @@
-// Names are automatically added through ScreenIdGenerator.cs, deletions are done manually :)
-namespace VioletUI {
-	public enum ScreenId {
-		None = 0,
-	}
+//Names are automatically added through ScreenIdGenerator.cs, deletions are done manually in Assets/Plugins/VioletUI/ScreenIds.json :)
+public enum ScreenId {
+	None = 0,
+	ScreenA = 1,
+	ScreenB = 2,
+	Showhide = 3,
 }